idle_shutdown/requirements.txt
google-api-python-client
google-cloud-monitoring
